created anisotropy utils
