+++ burnman/utils/__init__.py
@@ -23,3 +23,4 @@ from . import misc
 from . import reductions
 from . import unitcell
 from . import geotherm
+from . import anisotropy
