+++ examples/example_anisotropic_mineral.py
@@ -85,21 +85,21 @@ if __name__ == "__main__":
     T = 1600.0
     m.set_state(P, T)
     print(f"Model forsterite properties at {P/1.e9:.2f} GPa and {T:.2f} K:")
-    np.set_printoptions(precision=4)
+    np.set_printoptions(formatter={"all": lambda x: f"{x:0.4f}"})
     print("Cell vectors:")
     print(m.cell_vectors)
     print("Cell parameters:")
     print(m.cell_parameters)
-    print("Thermal expansivity:")
-    print(m.thermal_expansivity_tensor)
-    print("Isothermal compressibility:")
-    print(m.isothermal_compressibility_tensor)
-    print("Thermal stress tensor:")
-    print(m.thermal_stress_tensor)
-    print("Isothermal stiffness_tensor:")
-    print(m.isothermal_stiffness_tensor)
-    print("Isentropic stiffness_tensor:")
-    print(m.isentropic_stiffness_tensor)
+    print("Thermal expansivity (1/MK):")
+    print(m.thermal_expansivity_tensor * 1.0e6)
+    print("Isothermal compressibility (1/GPa):")
+    print(m.isothermal_compressibility_tensor * 1.0e9)
+    print("Thermal stress tensor (MPa/K):")
+    print(m.thermal_stress_tensor / 1.0e6)
+    print("Isothermal stiffness_tensor (GPa):")
+    print(m.isothermal_stiffness_tensor / 1.0e9)
+    print("Isentropic stiffness_tensor (GPa):")
+    print(m.isentropic_stiffness_tensor / 1.0e9)
     print("Grueneisen tensor:")
     print(m.grueneisen_tensor)
