+++ misc/ref/example_anisotropic_mineral.py.out
@@ -1,40 +1,40 @@
 Model forsterite properties at 1.00 GPa and 1600.00 K:
 Cell vectors:
-[[0.0255 0.     0.    ]
- [0.     0.0551 0.    ]
- [0.     0.     0.0322]]
+[[0.0255 0.0000 0.0000]
+ [0.0000 0.0551 0.0000]
+ [0.0000 0.0000 0.0322]]
 Cell parameters:
-[2.5464e-02 5.5083e-02 3.2167e-02 9.0000e+01 9.0000e+01 9.0000e+01]
-Thermal expansivity:
-[[8.6707e-06 0.0000e+00 0.0000e+00]
- [0.0000e+00 1.7341e-05 0.0000e+00]
- [0.0000e+00 0.0000e+00 1.3400e-05]]
-Isothermal compressibility:
-[[2.1212e-12 0.0000e+00 0.0000e+00]
- [0.0000e+00 4.2424e-12 0.0000e+00]
- [0.0000e+00 0.0000e+00 3.2782e-12]]
-Thermal stress tensor:
-[[-4087680.0929       -0.           -0.    ]
- [      -0.     -4087680.0929       -0.    ]
- [      -0.           -0.     -4087680.0929]]
-Isothermal stiffness_tensor:
-[[2.6487e+11 5.7471e+10 5.9288e+10 0.0000e+00 0.0000e+00 0.0000e+00]
- [5.7471e+10 1.5924e+11 6.1787e+10 0.0000e+00 0.0000e+00 0.0000e+00]
- [5.9288e+10 6.1787e+10 1.8672e+11 0.0000e+00 0.0000e+00 0.0000e+00]
- [0.0000e+00 0.0000e+00 0.0000e+00 5.2647e+10 0.0000e+00 0.0000e+00]
- [0.0000e+00 0.0000e+00 0.0000e+00 0.0000e+00 6.4420e+10 0.0000e+00]
- [0.0000e+00 0.0000e+00 0.0000e+00 0.0000e+00 0.0000e+00 6.6913e+10]]
-Isentropic stiffness_tensor:
-[[2.7186e+11 6.4462e+10 6.6279e+10 0.0000e+00 0.0000e+00 0.0000e+00]
- [6.4462e+10 1.6623e+11 6.8778e+10 0.0000e+00 0.0000e+00 0.0000e+00]
- [6.6279e+10 6.8778e+10 1.9371e+11 0.0000e+00 0.0000e+00 0.0000e+00]
- [0.0000e+00 0.0000e+00 0.0000e+00 5.2647e+10 0.0000e+00 0.0000e+00]
- [0.0000e+00 0.0000e+00 0.0000e+00 0.0000e+00 6.4420e+10 0.0000e+00]
- [0.0000e+00 0.0000e+00 0.0000e+00 0.0000e+00 0.0000e+00 6.6913e+10]]
+[0.0255 0.0551 0.0322 90.0000 90.0000 90.0000]
+Thermal expansivity (1/MK):
+[[8.6707 0.0000 0.0000]
+ [0.0000 17.3415 0.0000]
+ [0.0000 0.0000 13.4002]]
+Isothermal compressibility (1/GPa):
+[[0.0021 0.0000 0.0000]
+ [0.0000 0.0042 0.0000]
+ [0.0000 0.0000 0.0033]]
+Thermal stress tensor (MPa/K):
+[[-4.0877 -0.0000 -0.0000]
+ [-0.0000 -4.0877 -0.0000]
+ [-0.0000 -0.0000 -4.0877]]
+Isothermal stiffness_tensor (GPa):
+[[264.8653 57.4708 59.2880 0.0000 0.0000 0.0000]
+ [57.4708 159.2372 61.7868 0.0000 0.0000 0.0000]
+ [59.2880 61.7868 186.7232 0.0000 0.0000 0.0000]
+ [0.0000 0.0000 0.0000 52.6474 0.0000 0.0000]
+ [0.0000 0.0000 0.0000 0.0000 64.4195 0.0000]
+ [0.0000 0.0000 0.0000 0.0000 0.0000 66.9132]]
+Isentropic stiffness_tensor (GPa):
+[[271.8561 64.4616 66.2789 0.0000 0.0000 0.0000]
+ [64.4616 166.2280 68.7776 0.0000 0.0000 0.0000]
+ [66.2789 68.7776 193.7140 0.0000 0.0000 0.0000]
+ [0.0000 0.0000 0.0000 52.6474 0.0000 0.0000]
+ [0.0000 0.0000 0.0000 0.0000 64.4195 0.0000]
+ [0.0000 0.0000 0.0000 0.0000 0.0000 66.9132]]
 Grueneisen tensor:
-[[1.0689 0.     0.    ]
- [0.     1.0689 0.    ]
- [0.     0.     1.0689]]
+[[1.0689 0.0000 0.0000]
+ [0.0000 1.0689 0.0000]
+ [0.0000 0.0000 1.0689]]
 Volume: 45.1175 cm^3/mol
 Entropy: 359.02 J/K/mol
 C_P: 184.17 J/K/mol
